Add default value for second argument of fetchChildByName, that enables it to be used conveniently on C strings. Add copyName helper into Node, to ease its use with pure C strings (via some kind of utility decorator CRTP stuff). Check with -m32 and -m64 -Wall proofing all tests

+++ test/MockStorage.h
@@ -30,6 +30,7 @@
 #include <cstring>
 #include <unordered_set>
 #include <map>
+#include <cstdint>
 
 class MockStorageTrace;
 
@@ -92,7 +93,7 @@ struct MockStorage: protected ubiq::Trace<MockStorageTrace> {
 	// These are the required definitions mentioned above
 	////////////////////////////////////////////////////////////////
 
-	typedef unsigned int Address;
+	typedef uintptr_t Address;
 	static const Address InvalidAddress;
 	static const unsigned int pageSize = pageSizeParam;
